Aprimoramento geral
Métodos principais de "Endereco" e "Contato" só são chamados pelos
métodos de objetos de classes herdadas; métodos de "Endereco" e
"Contato" retornam para os métodos que os chamam finalizarem em caso de
erro; verificações agora são realizadas dentro dos métodos; retirado
código desnecessário para inserir produto no estoque direto da remessa;
"db-queries.php" excluído; "funcoesBase.php" guarda as funções comuns
aos arquivos; "manager.php" com redução do código devido aos
aprimoramentos anteriores; outras alterações minoritárias.

+++ php/class/Cliente.php
@@ -11,6 +11,7 @@ class Cliente extends Contato{
 		$this->obs=$obs;
 	}
 	public function cadastrarCliente(){
+		if($this->cadastrarEndereco()===false||$this->cadastrarContato()===false){return;}
 		$mysqli=$this->conectar();
 		$queryInsert='insert into cliente(nome,cpf,obs,endereco,contato) values ("'.$this->nome.'","'.$this->cpf.'","'.$this->obs.'",'.$this->idEndereco.','.$this->idContato.');';
 		if(!mysqli_query($mysqli,$queryInsert)){
@@ -21,7 +22,7 @@ class Cliente extends Contato{
 		}
 	}
 	public function buscarDadosCliente(){
-		$mysqli=$this->conectar();
+		if($this->verificarExistencia('cliente','id',$this->idCliente)===false){return;}
 		$this->nome=$this->getValueInBank('nome','cliente','id',$this->idCliente);
 		$this->cpf=$this->getValueInBank('cpf','cliente','id',$this->idCliente);
 		$this->obs=$this->getValueInBank('obs','cliente','id',$this->idCliente);
@@ -42,6 +43,7 @@ class Cliente extends Contato{
 		$mysqli=$this->conectar();
 		$this->idContato=$this->getValueInBank('contato','cliente','id',$this->idCliente);
 		$this->idEndereco=$this->getValueInBank('endereco','cliente','id',$this->idCliente);
+		if($this->atualizarEndereco()===false||$this->atualizarContato()===false){return;}
 		$updCliente='update cliente set nome="'.$this->nome.'",cpf="'.$this->cpf.'",obs="'.$this->obs.'" where id='.$this->idCliente.';';
 		if(!mysqli_query($mysqli,$updCliente)){
 			die ('<script>alert("Não foi possível atualizar o cliente:\n\n'.mysqli_error($mysqli).'");location.href="/trabalhos/gti/bda1/";</script>');
@@ -50,11 +52,11 @@ class Cliente extends Contato{
 		}
 	}
 	public function excluirCliente(){
+		if($this->verificarExistencia('cliente','id',$this->idCliente)===false){return;}
 		$this->nome=$this->getValueInBank('nome','cliente','id',$this->idCliente);
 		$this->idContato=$this->getValueInBank('contato','cliente','id',$this->idCliente);
 		$this->idEndereco=$this->getValueInBank('endereco','cliente','id',$this->idCliente);
-		$this->excluirContato();
-		$this->excluirEndereco();
+		if($this->excluirEndereco()===false||$this->excluirContato()===false){return;}
 		$delCliente='delete from cliente where id='.$this->idCliente.';';
 		$mysqli=$this->conectar();
 		if(!mysqli_query($mysqli,$delCliente)){
